fix(hentaihand/nhentaicom): fix lookup filters parsing numeric ids (#18181)
fix(hentaihand): fix lookup filters parsing numeric ids

+++ lib-multisrc/hentaihand/src/eu/kanade/tachiyomi/multisrc/hentaihand/HentaiHand.kt
@@ -96,7 +96,7 @@ abstract class HentaiHand :
                 if (idList.isEmpty()) {
                     return@map null
                 } else {
-                    idList.first().toInt()
+                    idList.first()
                 }
             }.toBlocking().first()
     }
